requirements.txt
blinker==1.7.0
openai==1.14.2
pydantic==2.6.4
Pygments==2.16.1
python-dotenv==1.0.1
Requests==2.31.0
rich==13.7.1
streamlit==1.32.2
